+++ src/Exceptions/ListStoreException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Crwlr\Url\Exceptions;
+
+use Exception;
+
+class ListStoreException extends Exception
+{
+}